requirements.txt
django
django-environ
django-axes
reportlab
pyotp
qrcode
Pillow